Working version selector - awaiting dropdown component
