+++ package.json
@@ -82,6 +82,7 @@
     "mongodb-stitch-server-sdk": "^4.7.1",
     "no-scroll": "^2.1.1",
     "prop-types": "^15.7.2",
+    "query-string": "^6.13.1",
     "rasterizehtml": "^1.3.0",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
