+++ src/components/ComponentFactory.js
@@ -50,7 +50,6 @@ import Image from './Image';
 import RefRole from './RefRole';
 import Target from './Target';
 import Glossary from './Glossary';
-import DeprecatedVersionSelector from './DeprecatedVersionSelector';
 
 import RoleAbbr from './Roles/Abbr';
 import RoleClass from './Roles/Class';
@@ -93,7 +92,6 @@ export default class ComponentFactory extends Component {
       definitionList: DefinitionList,
       definitionListItem: DefinitionListItem,
       deprecated: Deprecated,
-      'deprecated-version-selector': DeprecatedVersionSelector,
       emphasis: Emphasis,
       figure: Figure,
       footnote: Footnote,
