+++ src/components/landing/ComponentFactory.js
@@ -3,12 +3,14 @@ import PropTypes from 'prop-types';
 import Card from './Card';
 import CardGroup from './CardGroup';
 import CTA from './CTA';
+import DeprecatedVersionSelector from './DeprecatedVersionSelector';
 import Introduction from './Introduction';
 
 const componentMap = {
   card: Card,
   'card-group': CardGroup,
   cta: CTA,
+  'deprecated-version-selector': DeprecatedVersionSelector,
   introduction: Introduction,
 };
